+++ ai/backend/client/auth.py
@@ -16,7 +16,7 @@ def generate_signature(method, version, endpoint,
     body_hash = hashlib.new(hash_type, content).hexdigest()
     major_ver = version.split('.', 1)[0]
 
-    sign_str = '{}\n/{}/{}\n{}\nhost:{}\ncontent-type:{}\nx-sorna-version:{}\n{}'.format(  # noqa
+    sign_str = '{}\n/{}/{}\n{}\nhost:{}\ncontent-type:{}\nx-backendai-version:{}\n{}'.format(  # noqa
         method.upper(),
         major_ver, request_path,
         date.isoformat(),
@@ -33,7 +33,7 @@ def generate_signature(method, version, endpoint,
 
     signature = hmac.new(sign_key, sign_bytes, hash_type).hexdigest()
     headers = {
-        'Authorization': 'Sorna signMethod=HMAC-{}, credential={}:{}'.format(
+        'Authorization': 'BackendAI signMethod=HMAC-{}, credential={}:{}'.format(
             hash_type.upper(),
             access_key,
             signature