+++ ai/backend/client/cli/__init__.py
@@ -43,7 +43,7 @@ def register_command(handler: Callable[[argparse.Namespace], None],
 
 def main():
 
-    import sorna.cli.run  # noqa
+    import ai.backend.client.cli.run  # noqa
 
     mode = Path(sys.argv[0]).stem