+++ tests/test_config.py
@@ -1,6 +1,6 @@
 import pytest
 
-from sorna.config import APIConfig, get_config, set_config
+from ai.backend.client.config import APIConfig, get_config, set_config
 
 
 @pytest.fixture
@@ -8,7 +8,7 @@ def cfg_params():
     return {
         'endpoint': 'http://127.0.0.1:8081',
         'version': 'vtest',
-        'user_agent': 'Sorna Client Test',
+        'user_agent': 'Backed.AI Client Test',
         'access_key': 'AKIAIOSFODNN7EXAMPLE',
         'secret_key': 'wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY',
         'hash_type': 'md5',
@@ -37,7 +37,7 @@ def test_api_config_initialization(cfg_params):
 def test_set_and_get_config(mocker, cfg_params):
     # Mocking the global variable ``_config``.
     # The value of a global variable will affect other test cases.
-    mocker.patch('sorna.config._config', None)
+    mocker.patch('ai.backend.client.config._config', None)
 
     cfg = APIConfig(**cfg_params)
     set_config(cfg)
@@ -47,7 +47,7 @@ def test_set_and_get_config(mocker, cfg_params):
 
 def test_access_and_secret_key_should_be_set_to_get_default_config(
         mocker, cfg_params):
-    mocker.patch('sorna.config._config', None)
+    mocker.patch('ai.backend.client.config._config', None)
 
     # Neither SORNA_ACCESS_KEY nor SORNA_SECRET_KEY exists.
     mocker.patch('os.environ', {})
@@ -78,7 +78,7 @@ def test_access_and_secret_key_should_be_set_to_get_default_config(
 
 def test_get_config_return_default_config_when_config_is_none(
         mocker, cfg_params):
-    mocker.patch('sorna.config._config', None)
+    mocker.patch('ai.backend.client.config._config', None)
     mocker.patch('os.environ', {
         'SORNA_ACCESS_KEY': cfg_params['access_key'],
         'SORNA_SECRET_KEY': cfg_params['secret_key']