+++ tests/test_kernel.py
@@ -2,9 +2,9 @@ from unittest import mock
 
 import pytest
 
-from sorna.exceptions import SornaAPIError
-from sorna.compat import token_hex
-from sorna.kernel import (
+from ai.backend.client.exceptions import BackendAPIError
+from ai.backend.client.compat import token_hex
+from ai.backend.client.kernel import (
     create_kernel, destroy_kernel, restart_kernel, get_kernel_info,
     execute_code
 )
@@ -14,7 +14,7 @@ def test_create_kernel_url(mocker):
     mock_req_obj = mock.Mock()
     mock_req_obj.send.return_value = mocker.MagicMock(status=201,
                                                       json=mock.MagicMock())
-    mock_req = mocker.patch('sorna.kernel.Request', return_value=mock_req_obj)
+    mock_req = mocker.patch('ai.backend.client.kernel.Request', return_value=mock_req_obj)
 
     create_kernel('python')
 
@@ -29,7 +29,7 @@ def test_create_kernel_return_id_only(mocker):
     mock_req_obj.send.return_value = mock.MagicMock(
         status=201, json=mock_json_func
     )
-    mocker.patch('sorna.kernel.Request', return_value=mock_req_obj)
+    mocker.patch('ai.backend.client.kernel.Request', return_value=mock_req_obj)
 
     k = create_kernel('python')
 
@@ -39,16 +39,16 @@ def test_create_kernel_return_id_only(mocker):
 def test_create_kernel_raises_err_with_abnormal_status(mocker):
     mock_req_obj = mock.Mock()
     mock_req_obj.send.return_value = mock.MagicMock(status=400)
-    mocker.patch('sorna.kernel.Request', return_value=mock_req_obj)
+    mocker.patch('ai.backend.client.kernel.Request', return_value=mock_req_obj)
 
-    with pytest.raises(SornaAPIError):
+    with pytest.raises(BackendAPIError):
         create_kernel('python')
 
 
 def test_destroy_kernel_url(mocker):
     mock_req_obj = mock.Mock()
     mock_req_obj.send.return_value = mock.MagicMock(status=204)
-    mock_req = mocker.patch('sorna.kernel.Request', return_value=mock_req_obj)
+    mock_req = mocker.patch('ai.backend.client.kernel.Request', return_value=mock_req_obj)
 
     kernel_id = token_hex(12)
     destroy_kernel(kernel_id)
@@ -60,17 +60,17 @@ def test_destroy_kernel_url(mocker):
 def test_destroy_kernel_raises_err_with_abnormal_status(mocker):
     mock_req_obj = mock.Mock()
     mock_req_obj.send.return_value = mock.MagicMock(status=400)
-    mocker.patch('sorna.kernel.Request', return_value=mock_req_obj)
+    mocker.patch('ai.backend.client.kernel.Request', return_value=mock_req_obj)
 
     kernel_id = token_hex(12)
-    with pytest.raises(SornaAPIError):
+    with pytest.raises(BackendAPIError):
         destroy_kernel(kernel_id)
 
 
 def test_restart_kernel_url(mocker):
     mock_req_obj = mock.Mock()
     mock_req_obj.send.return_value = mock.MagicMock(status=204)
-    mock_req = mocker.patch('sorna.kernel.Request', return_value=mock_req_obj)
+    mock_req = mocker.patch('ai.backend.client.kernel.Request', return_value=mock_req_obj)
 
     kernel_id = token_hex(12)
     restart_kernel(kernel_id)
@@ -82,17 +82,17 @@ def test_restart_kernel_url(mocker):
 def test_restart_kernel_raises_err_with_abnormal_status(mocker):
     mock_req_obj = mock.Mock()
     mock_req_obj.send.return_value = mock.MagicMock(status=400)
-    mocker.patch('sorna.kernel.Request', return_value=mock_req_obj)
+    mocker.patch('ai.backend.client.kernel.Request', return_value=mock_req_obj)
 
     kernel_id = token_hex(12)
-    with pytest.raises(SornaAPIError):
+    with pytest.raises(BackendAPIError):
         restart_kernel(kernel_id)
 
 
 def test_get_kernel_info_url(mocker):
     mock_req_obj = mock.Mock()
     mock_req_obj.send.return_value = mock.MagicMock(status=200)
-    mock_req = mocker.patch('sorna.kernel.Request', return_value=mock_req_obj)
+    mock_req = mocker.patch('ai.backend.client.kernel.Request', return_value=mock_req_obj)
 
     kernel_id = token_hex(12)
     get_kernel_info(kernel_id)
@@ -105,17 +105,17 @@ def test_get_kernel_info_url(mocker):
 def test_get_kernel_info_raises_err_with_abnormal_status(mocker):
     mock_req_obj = mock.Mock()
     mock_req_obj.send.return_value = mock.MagicMock(status=400)
-    mocker.patch('sorna.kernel.Request', return_value=mock_req_obj)
+    mocker.patch('ai.backend.client.kernel.Request', return_value=mock_req_obj)
 
     kernel_id = token_hex(12)
-    with pytest.raises(SornaAPIError):
+    with pytest.raises(BackendAPIError):
         get_kernel_info(kernel_id)
 
 
 def test_execute_code_url(mocker):
     mock_req_obj = mock.Mock()
     mock_req_obj.send.return_value = mock.MagicMock(status=200)
-    mock_req = mocker.patch('sorna.kernel.Request', return_value=mock_req_obj)
+    mock_req = mocker.patch('ai.backend.client.kernel.Request', return_value=mock_req_obj)
 
     kernel_id = token_hex(12)
     execute_code(kernel_id, 'hello')
@@ -129,8 +129,8 @@ def test_execute_code_url(mocker):
 def test_execute_code_raises_err_with_abnormal_status(mocker):
     mock_req_obj = mock.Mock()
     mock_req_obj.send.return_value = mock.MagicMock(status=400)
-    mocker.patch('sorna.kernel.Request', return_value=mock_req_obj)
+    mocker.patch('ai.backend.client.kernel.Request', return_value=mock_req_obj)
 
     kernel_id = token_hex(12)
-    with pytest.raises(SornaAPIError):
+    with pytest.raises(BackendAPIError):
         execute_code(kernel_id, 'hello')